Refactor Activity Timeline, updated other components for clarity
